:twisted_rightwards_arrows: Merge(feature/be/gameAPI리팩토링 -> develop): 게임결과로직수정

See merge request s09-final/S09P31A501!143

+++ backend/src/main/java/com/npc/say_vr/domain/game/domain/Ranking.java
@@ -33,7 +33,7 @@ public class Ranking {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private int point;
+    private Long point;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "tier_id")
